refactor(backend): rewrite CLI and remove global logger
Move away from Cobra + Viper for CLI and instead move to Kong. This
library allows CLI to be declared via structs instead of manually adding
flags, as well as removing the need for a separate configuration
library and global state. This allows components of the CLI to be tested
independently.

With this, we are also removing the global logger in favor of passing it
via contexts, which would allow for things like route-specific loggers.