backend: implement /healthz route
This route allows a service manager (ie. Docker) to monitor whether the
server is ready to serve or is still functional, which is important for
features like "zero-downtime" upgrades.

In addition to this route, a simple client for running the health check
is added to parkserver CLI. This allow us to remove the need for HTTP
clients inside the server container.