ci(backend): build container and deploy to render.com (#180)
For PR, we build container in a similar manner to the deployment
workflow in main. Additionally, upload the result to artifacts so
reviewers can download and load into their own Docker instance for
testing.

For push to main or tags, build the backend container and upload to
GitHub Package Registry. Then, trigger deployment to staging or
production depending on whether the push is to main or to a release.