fix: Delaying converting to UTC
We will convert to UTC time upon before making request only. Doing editing in UTC time is not
intuitive to user so we will use local time zone for editing and convert that to UTC before sending it to the backend.
For the record, when receiving time slot from backend, we will also convert it to local first before displaying