android: centralize cache creation on main (#220)
Apparently cache are not shared between branches at all, even if that
branch was merged to main. Only caches that are created on main can be
preserved.

This makes Android CI extremely wasteful due to the emulator image being
~2GiB in size. Remedy this problem by making the CI restore-only, and
populate the cache whenever an android change hits main.

We should see a speed up for Android CI following this.